Move SportLeague to a common location so that it's more easily used across modules
