+++ backend/src/common/types/sport-league.type.ts
@@ -1,0 +1,1 @@
+export type SportLeague = 'NFL' | 'GOLF' | 'NBA' | 'NHL' | 'MLB' | 'WORLDCUP';
