+++ backend/src/leagues/entities/league.entity.ts
@@ -1,5 +1,6 @@
 import { Selectable } from 'kysely';
 import { tags } from 'typia';
+import { SportLeague } from '@/common/types/sport-league.type';
 
 export interface ILeague {
   leagueId: string & tags.Format<'uuid'>;
@@ -10,4 +11,3 @@ export interface ILeague {
 export type League = Selectable<ILeague>;
 export type CreateLeagueDto = Omit<ILeague, 'leagueId'>;
 export type UpdateLeagueDto = Partial<ILeague>;
-export type SportLeague = 'NFL' | 'GOLF' | 'NBA' | 'NHL' | 'MLB' | 'WORLDCUP';
